Shortened Model 100 epochs
This time I shortened the model to 6k lines only, and changed the runtime from T4 to normal CPU. I did this because I had errors from the google collab console that the T4 GPU wasn't being used, and that I should switch to a different run time. It was much faster this time.